Initial GUI Development
Provided is the initial GUI set up with little to no polish. Focuses on formatting the proper organization of the Game and End screens. Will later include cleaner formatting and usable parts.

The switch button at the bottom allows you to switch between the two availaible screens, this is temporary and serves only to see both screens.

The program is developed with the intent that a single JFrame/window is made and simply modified.

With this in mind, the existing GameOverScreen will be null/unnecessary as this push incorporates an End screen.